+++ src/getters.jl
@@ -1175,7 +1175,7 @@ function GetVisitDate(
 end
 
 """
-function function GetVisitDate(df:DataFrame, conn; interval::Symbol = :start, tab = visit_occurrence)
+function GetVisitDate(df:DataFrame, conn; interval::Symbol = :start, tab = visit_occurrence)
 
 Given a `DataFrame` with a `:visit_occurrence_id` column, return the `visit_occurrence_id` and either `visit_start_date` or `visit_end_date`, depending on the value of the `interval` for each `visit_occurrence_id` 
 
@@ -1238,7 +1238,7 @@ function GetVisitDate(
 end
 
 """
-GetDrugExposures(ids, conn; tab = drug_exposure)
+GetDrugExposureIDs(ids, conn; tab = drug_exposure)
 
 Given a list of person IDs, find their drug exposure.
 
@@ -1256,26 +1256,26 @@ Given a list of person IDs, find their drug exposure.
 
 - `df::DataFrame` - a two column `DataFrame` comprised of columns: `:person_id` and `:drug_exposure_id`
 """
-function GetDrugExposures(
+function GetDrugExposureIDs(
     ids,
     conn;
     tab=drug_exposure
 )
-    df = DBInterface.execute(conn, GetDrugExposures(ids; tab=tab)) |> DataFrame
+    df = DBInterface.execute(conn, GetDrugExposureIDs(ids; tab=tab)) |> DataFrame
 
     return df
 
 end
 
 """
-function GetDrugExposures(df:DataFrame, conn; tab = drug_exposure)
+function GetDrugExposureIDs(df:DataFrame, conn; tab = drug_exposure)
 
 Given a `DataFrame` with a `:person_id` column, return the `DataFrame` with an associated `:drug_exposure_id`for each `person_id` in the `DataFrame`
 
-Multiple dispatch that accepts all other arguments like in `GetDrugExposures(ids, conn; tab = drug_exposure)`
+Multiple dispatch that accepts all other arguments like in `GetDrugExposureIDs(ids, conn; tab = drug_exposure)`
 """
 
-function GetDrugExposures(
+function GetDrugExposureIDs(
     df::DataFrame,
     conn;
     tab=drug_exposure
@@ -1284,11 +1284,11 @@ function GetDrugExposures(
     df_ids= df[:,"person_id"]
     
 
-    return outerjoin(GetDrugExposures(df_ids, conn; tab=tab), df, on = :person_id)
+    return outerjoin(GetDrugExposureIDs(df_ids, conn; tab=tab), df, on = :person_id)
 end
 
 """
-GetDrugExposures(ids; tab = drug_exposure)
+GetDrugExposureIDs(ids; tab = drug_exposure)
 
 Return SQL statement that gets the `drug_exposure_id` for a given list of `person_id`'s
 
@@ -1305,7 +1305,7 @@ Return SQL statement that gets the `drug_exposure_id` for a given list of `perso
 - `df::DataFrame` - a two column `DataFrame` comprised of columns: `:person_id` and `:drug_exposure_id`
 """
 
-function GetDrugExposures(
+function GetDrugExposureIDs(
     ids;
     tab=drug_exposure
 )
@@ -1481,4 +1481,408 @@ function GetDrugAmounts(
 
 end
 
-export GetDatabasePersonIDs, GetPatientState, GetPatientGender, GetPatientRace, GetPatientAgeGroup, GetPatientVisits, GetMostRecentConditions, GetMostRecentVisit, GetVisitCondition, GetPatientEthnicity, GetDatabaseYearRange, GetVisitPlaceOfService, GetVisitConcept, GetVisitDate, GetDrugExposures, GetDrugConceptIDs, GetDrugAmounts
+"""
+GetVisitProcedure(visit_ids, conn; tab = procedure_occurrence)
+
+Given a list of visit IDs, find their corresponding procedures.
+
+# Arguments:
+
+- `visit_ids` - list of `visit_id`'s; each ID must be of subtype `Integer`
+
+- `conn` - database connection using DBInterface
+
+# Keyword Arguments:
+
+- `tab` - the `SQLTable` representing the Condition Occurrence table; default `procedure_occurrence`
+
+# Returns
+
+- `df::DataFrame` - a two column `DataFrame` comprised of columns: `:visit_occurrence_id` and `:procedure_concept_id`
+"""
+function GetVisitProcedure(
+    visit_ids,
+    conn;
+    tab=procedure_occurrence
+)
+
+    df = DBInterface.execute(conn, GetVisitProcedure(visit_ids; tab=tab)) |> DataFrame
+
+    return df
+
+end
+
+"""
+function GetVisitProcedure(df:DataFrame, conn; tab = procedure_occurrence)
+
+Given a `DataFrame` with a `:visit_occurrence_id` column, return the `DataFrame` with an associated `:procedure_concept_id` for each `visit_occurrence_id` in the `DataFrame`
+
+Multiple dispatch that accepts all other arguments like in `GetVisitProcedure(ids, conn; tab = procedure_occurrence)`
+"""
+
+function GetVisitProcedure(
+    df::DataFrame,
+    conn;
+    tab=procedure_occurrence
+)
+
+    df_ids= df[:,"visit_occurrence_id"]
+    
+    return outerjoin(GetVisitProcedure(df_ids, conn; tab=tab), df, on = :visit_occurrence_id)
+
+end
+
+"""
+GetVisitProcedure(visit_ids; tab = procedure_occurrence)
+
+Produces SQL statement that, given a list of `visit_id`'s, finds the procedures associated with that visit.
+
+# Arguments:
+
+- `visit_ids` - list of `visit_id`'s; each ID must be of subtype `Integer`
+
+# Keyword Arguments:
+
+- `tab` - the `SQLTable` representing the Procedure Occurrence table; default `procedure_occurrence`
+
+# Returns
+
+- `df::DataFrame` - a two column `DataFrame` comprised of columns: `:visit_occurrence_id` and `:procedure_concept_id`
+"""
+function GetVisitProcedure(
+    visit_ids;
+    tab=procedure_occurrence
+)
+
+    sql =
+        From(tab) |>
+        Where(Fun.in(Get.visit_occurrence_id, visit_ids...)) |>
+        Select(Get.visit_occurrence_id, Get.procedure_concept_id) |>
+        q -> render(q, dialect=dialect)
+
+    return String(sql)
+
+end
+
+"""
+GetCohortSubjects(cohort_ids, conn; tab = cohort)
+
+Given a list of cohort IDs, find their corresponding subjects.
+
+# Arguments:
+
+- `cohort_ids` - list of `cohort_id`'s; each ID must be of subtype `Float64`
+
+- `conn` - database connection using DBInterface
+
+# Keyword Arguments:
+
+- `tab` - the `SQLTable` representing the `cohort` table; default `cohort`
+
+# Returns
+
+- `df::DataFrame` - a two column `DataFrame` comprised of columns: `:cohort_definition_id` and `:subject_id`
+"""
+function GetCohortSubjects(
+    cohort_ids, 
+    conn; 
+    tab = cohort
+)
+
+    df = DBInterface.execute(conn, GetCohortSubjects(cohort_ids; tab=tab)) |> DataFrame
+
+    return df
+
+end
+
+"""
+function GetCohortSubjects(df:DataFrame, conn; tab = cohort)
+
+Given a `DataFrame` with a `:cohort_definition_id` column, return the `DataFrame` with an associated `:subject_id` for each `cohort_definition_id` in the `DataFrame`
+
+Multiple dispatch that accepts all other arguments like in `GetCohortSubjects(ids, conn; tab = cohort)`
+"""
+
+function GetCohortSubjects(
+    df::DataFrame,
+    conn;
+    tab = cohort
+)
+
+    df_ids= df[:,"cohort_definition_id"]
+    
+    return outerjoin(GetCohortSubjects(df_ids, conn; tab=tab), df, on = :cohort_definition_id)
+
+end
+
+"""
+GetCohortSubjects(cohort_ids; tab = cohort)
+
+Produces SQL statement that, given a list of `cohort_id`'s, finds the subjects associated with that cohort.
+
+# Arguments:
+
+- `cohort_ids` - list of `cohort_id`'s; each ID must be of subtype `Float64`
+
+# Keyword Arguments:
+
+- `tab` - the `SQLTable` representing the `cohort` table; default `cohort`
+
+# Returns
+
+- `df::DataFrame` - a two column `DataFrame` comprised of columns: `:cohort_definition_id` and `:subject_id`
+"""
+function GetCohortSubjects(
+    cohort_ids;
+    tab = cohort
+)
+
+    sql =
+        From(tab) |>
+        Where(Fun.in(Get.cohort_definition_id, cohort_ids...)) |>
+        Select(Get.cohort_definition_id, Get.subject_id) |>
+        q -> render(q, dialect=dialect)
+
+    return String(sql)
+
+end
+
+"""
+function GetCohortSubjectStartDate(cohort_ids, subject_ids, conn; tab=cohort)
+
+    Given a single or list of cohort IDs and subject IDs, return their start dates.
+    
+    # Arguments:
+    
+    - `cohort_ids` - list of `cohort_id`'s; each ID must be of subtype `Float64`
+        
+    - `subject_id` - list of `subject_id`'s; each ID must be of subtype `Float64`
+        
+    - `conn` - database connection using DBInterface
+        
+    # Keyword Arguments:
+        
+    - `tab` - the `SQLTable` representing the `cohort` table; default `cohort`
+
+# Returns
+
+- `df::DataFrame` - a three column `DataFrame` comprised of columns: `:cohort_definition_id` , `:subject_id` and `:cohort_start_date`
+"""
+function GetCohortSubjectStartDate(
+    cohort_ids,
+    subject_ids,
+    conn;
+    tab = cohort
+)
+
+    df = DBInterface.execute(conn, GetCohortSubjectStartDate(cohort_ids, subject_ids; tab=tab)) |> DataFrame
+
+    return df
+    
+end
+
+"""
+function GetCohortSubjectStartDate(df:DataFrame, conn; tab = cohort)
+
+Given a `DataFrame` with a `:cohort_definition_id` column and `:subject_id` column, return the `DataFrame` with an associated `:cohort_start_date` corresponding to a cohort's subject ID in the `DataFrame`
+
+Multiple dispatch that accepts all other arguments like in `GetCohortSubjectStartDate(ids, conn; tab = cohort)`
+"""
+function GetCohortSubjectStartDate(
+    df::DataFrame, 
+    conn; 
+    tab = cohort
+)
+
+    return outerjoin(GetCohortSubjectStartDate(df[:,"cohort_definition_id"], df[:,"subject_id"], conn; tab=tab), df, on = :cohort_definition_id)
+
+end
+
+"""
+function GetCohortSubjectStartDate(cohort_ids, subject_ids; tab=cohort)
+
+Given a list of cohort IDs and subject IDs return their start dates.
+
+# Arguments:
+
+- `cohort_ids` - list of `cohort_id`'s; each ID must be of subtype `Float64`
+    
+- `subject_id` - list of `subject_id`'s; each ID must be of subtype `Float64`
+    
+- `conn` - database connection using DBInterface
+    
+# Keyword Arguments:
+    
+- `tab` - the `SQLTable` representing the `cohort` table; default `cohort`
+    
+# Returns
+    
+- `df::DataFrame` - a three column `DataFrame` comprised of columns: `:cohort_definition_id` , `:subject_id` and `:cohort_start_date`
+
+"""
+function GetCohortSubjectStartDate(
+    cohort_ids,
+    subject_ids;
+    tab = cohort
+)
+
+    sql =
+        From(tab) |>
+        Where(Fun.in(Get.cohort_definition_id, cohort_ids...)) |>
+        Where(Fun.in(Get.subject_id, subject_ids...)) |>
+        Select(Get.cohort_definition_id, Get.subject_id, Get.cohort_start_date) |>
+        q -> render(q, dialect=dialect)
+
+    return String(sql)
+    
+end
+
+"""
+function GetCohortSubjectEndDate(cohort_ids, subject_ids, conn; tab=cohort)
+
+    Given a list of cohort IDs and subject IDs return their end dates.
+    
+    # Arguments:
+    
+    - `cohort_ids` - list of `cohort_id`'s; each ID must be of subtype `Float64`
+        
+    - `subject_id` - list of `subject_id`'s; each ID must be of subtype `Float64`
+        
+    - `conn` - database connection using DBInterface
+        
+    # Keyword Arguments:
+        
+    - `tab` - the `SQLTable` representing the `cohort` table; default `cohort`
+
+# Returns
+
+- `df::DataFrame` - a three column `DataFrame` comprised of columns: `:cohort_definition_id` , `:subject_id` and `:cohort_end_date`
+"""
+function GetCohortSubjectEndDate(
+    cohort_ids,
+    subject_ids,
+    conn;
+    tab = cohort
+)
+
+    df = DBInterface.execute(conn, GetCohortSubjectEndDate(cohort_ids, subject_ids; tab=tab)) |> DataFrame
+
+    return df
+    
+end
+
+"""
+function GetCohortSubjectEndDate(df:DataFrame, conn; tab = cohort)
+
+Given a `DataFrame` with a `:cohort_definition_id` column and `:subject_id` column, return the `DataFrame` with an associated `:cohort_end_date` corresponding to a given `cohort_definition_id` and `subject_id` in the `DataFrame`
+
+Multiple dispatch that accepts all other arguments like in `GetCohortSubjectEndDate(ids, conn; tab = cohort)`
+"""
+function GetCohortSubjectEndDate(
+    df::DataFrame, 
+    conn; 
+    tab = cohort
+)
+
+    return outerjoin(GetCohortSubjectEndDate(df[:,"cohort_definition_id"], df[:,"subject_id"], conn; tab=tab), df, on = :cohort_definition_id)
+
+end
+
+"""
+function GetCohortSubjectEndDate(cohort_ids; subject_ids; tab=cohort)
+
+Given a list of cohort IDs and subject IDs return their end date.
+
+# Arguments:
+
+- `cohort_ids` - list of `cohort_id`'s; each ID must be of subtype `Float64`
+    
+- `subject_id` - list of `subject_id`'s; each ID must be of subtype `Float64`
+    
+- `conn` - database connection using DBInterface
+    
+# Keyword Arguments:
+    
+- `tab` - the `SQLTable` representing the `cohort` table; default `cohort`
+   
+# Returns
+    
+- `df::DataFrame` - a three column `DataFrame` comprised of columns: `:cohort_definition_id` , `:subject_id` and `:cohort_end_date`
+
+"""
+function GetCohortSubjectEndDate(
+    cohort_ids,
+    subject_ids;
+    tab=cohort
+)
+
+    sql =
+        From(tab) |>
+        Where(Fun.in(Get.cohort_definition_id, cohort_ids...)) |>
+        Where(Fun.in(Get.subject_id, subject_ids...)) |>
+        Select(Get.cohort_definition_id, Get.subject_id, Get.cohort_end_date) |>
+        q -> render(q, dialect=dialect)
+
+    return String(sql)
+    
+end
+
+"""
+GetDatabaseCohorts(conn; tab=cohort)
+
+Given a `DataFrame` returns all unique cohort_definition_id associated with a database.
+
+#Arguments:
+
+- `conn` - database connection using DBInterface
+
+# Keyword Arguments:
+
+- `tab` - the `SQLTable` representing the Cohort table; default `cohort`
+
+# Returns
+    
+- `df::DataFrame` - a one column `DataFrame` comprised of columns: `:cohort_definition_id`
+
+
+"""
+function GetDatabaseCohorts(
+    conn; 
+    tab=cohort
+)
+    ids = DBInterface.execute(conn, GetDatabaseCohorts(tab=tab))  |> DataFrame
+
+    return convert(Vector{Int}, ids.cohort_definition_id)
+    
+end
+
+"""
+function GetDatabaseCohorts(; tab=cohort)
+
+Given a cohort table returns all unique IDs associated with a database.
+
+# Arguments:
+
+- `tab` - the `SQLTable` representing the Cohort table; default `cohort`
+
+# Returns
+    
+- `df::DataFrame` - a one column `DataFrame` comprised of columns: `:cohort_definition_id`
+
+"""
+
+function GetDatabaseCohorts(
+   ; tab=cohort
+)
+
+    sql = 
+        From(tab)  |>
+        Group(Get.cohort_definition_id) |>
+        Select(Get.cohort_definition_id) |>
+        q -> render(q, dialect=dialect)
+
+    return String(sql)
+    
+end
+
+export GetDatabasePersonIDs, GetPatientState, GetPatientGender, GetPatientRace, GetPatientAgeGroup, GetPatientVisits, GetMostRecentConditions, GetMostRecentVisit, GetVisitCondition, GetPatientEthnicity, GetDatabaseYearRange, GetVisitPlaceOfService, GetVisitConcept, GetVisitDate, GetDrugExposures, GetDrugConceptIDs, GetDrugAmounts, GetVisitProcedure, GetDatabaseCohorts, GetCohortSubjects, GetCohortSubjectStartDate, GetCohortSubjectEndDate, GetDrugExposureIDs